requirements.txt
requests
numpy
Pillow
torch